tests/wpt/tests/navigation-api/navigate-event/navigate-navigation-navigate.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
async_test(t => {
  navigation.onnavigate = t.step_func_done(e => {
    assert_equals(e.navigationType, "replace");
    assert_true(e.cancelable);
    assert_true(e.canIntercept);
    assert_false(e.userInitiated);
    assert_true(e.hashChange);
    assert_equals(e.downloadRequest, null);
    assert_equals(new URL(e.destination.url).hash, "#foo");
    assert_true(e.destination.sameDocument);
    assert_equals(e.destination.key, "");
    assert_equals(e.destination.id, "");
    assert_equals(e.destination.index, -1);
    assert_equals(e.formData, null);
    // NavigateEvent sourceElement is still in development, so test whether it is available.
    if ("sourceElement" in e) assert_equals(e.sourceElement, null);
  });
  navigation.navigate("#foo");
}, "navigate event for navigation.navigate()");
</script>
